% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToKoladaAPI.r
\name{flattenJSON}
\alias{flattenJSON}
\title{Flatten the JSON File}
\usage{
flattenJSON(df, df_data)
}
\arguments{
\item{df}{: level 1 flattened data from the verifContent function}

\item{df_data}{: Empty data frame for shuffling the data}
}
\value{
: Completely flattened json data as data.frame
}
\description{
Flatten the JSON File
}
\details{
: This function takes the data.frame with the list type and  flattens it into rows and column format
}
