% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToKoladaAPI.r
\name{getAPI}
\alias{getAPI}
\title{Get API data}
\usage{
getAPI(url)
}
\arguments{
\item{url}{: The URL to the API to fetch the data}
}
\value{
: Returns the object with details pertaining to accessing the API, contains the data and metadata
}
\description{
Get API data
}
\details{
:  This function tries to access the API from the input URL using the GET command httr package, and prints the status for the user.
            If the presented URL is invalid, the function throws the warning message and halts the execution
}
\examples{

\dontrun{
url = "http://api.kolada.se/v2/data/kpi/N01951,N61714/municipality/0180,0581,0580/year/2018,2019,2020"

response = getAPI(url)
GET URL STATUS: Success

}
}
