% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToKoladaAPI.r
\name{getKoladaAPIData}
\alias{getKoladaAPIData}
\title{User function to get the data from API}
\usage{
getKoladaAPIData(l_kpi = "", l_m = "", l_year = "", url = "")
}
\arguments{
\item{l_kpi}{: KPI code input ("code_kpi1, code_kpi1,..")to select specific KPIs from the API.}

\item{l_m}{: Municipality code input ("code_municipality1, code_municipality2,..") to get specific Municipality form the API}

\item{l_year}{: Year input ("year1, year2, ..."") for which the data to be observed}

\item{url}{: If data to be fetched directly from a specific URL}
}
\value{
The function returns a data.frame having the requested data from the API
}
\description{
User function to get the data from API
}
\details{
{This function to be called from the package to access the API, parse and download the data. Refer

\url{https://www.kolada.se/} for more details on Kolada API.

 Sameple KPIs codes and its description

N00002: Personalkostnader som andel av verksamhetens kostnader, andel (%)",

N00011: Inkomstutjämning, bidrag/avgift, kr/inv. 1 nov fg år",


 For more KPIs refer:
 \url{http://api.kolada.se/v2/kpi?all}

Sample Municipality Codes and values

 1480: Goteborg

 0580: Linkoping

 1281: Lund

 1280: Malmo


 For more municipalities refer:
 \url{http://api.kolada.se/v2/municipality?all}
 }
}
\examples{

\dontrun{

final_data = getKoladaAPIData(l_m = "0180,0581,0580",l_kpi = "N00042,N00019",l_year = "2018,2019,2020")

for more KPI codes
dplyr::select(getKoladaAPIData(url = "http://api.kolada.se/v2/kpi?all"),values.id, values.title)

for more Municipality codes
dplyr::select(getKoladaAPIData(url = "http://api.kolada.se/v2/municipality?all"),values.id, values.title)

 }

}
