% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToKoladaAPI.r
\name{getKoladaAPIData}
\alias{getKoladaAPIData}
\title{User function to get the data from API}
\usage{
getKoladaAPIData(l_kpi = "", l_m = "", l_year = "", url = "")
}
\arguments{
\item{l_kpi}{: Required KPI Code

 Sameple KPIs codes and its description

N00002: Personalkostnader som andel av verksamhetens kostnader, andel \(%\)",

N00011: Inkomstutjämning, bidrag/avgift, kr/inv. 1 nov fg år",

 For KPIs refer
 \url{http://api.kolada.se/v2/kpi?all}}

\item{l_m}{: Required Municipality Code

 Sample Municipality Codes and values

 1480: Goteborg

 0580: Linkoping

 1281: Lund

 1280: Malmo

 For more municipality refer
 \url{http://api.kolada.se/v2/municipality?all}}

\item{l_year}{: Required Year}

\item{url}{: If data to be fetched from a specific URL}
}
\value{
: data.frame having the requested data from the API
}
\description{
User function to get the data from API
}
\details{
: This function to be called from the package to access the API, parse and download the data.
}
\examples{

\dontrun{

final_data = getKoladaAPIData(l_m = "0180,0581,0580",l_kpi = "N00042,N00019",l_year = "2018,2019,2020")

for more KPI codes
dplyr::select(getKoladaAPIData(url = "http://api.kolada.se/v2/kpi?all"),values.id, values.title)

for more Municipality codes
 dplyr::select(getKoladaAPIData(url = "http://api.kolada.se/v2/municipality?all"),values.id, values.title)

 }

}
